+++ app/scripts/lforms-controllers.js
@@ -1189,4 +1189,6 @@ angular.module('lformsWidget')
             }
           }
         };
+
+
       }]);