+++ test/protractor/spec/lforms_testpage.po.js
@@ -434,7 +434,20 @@ var TestPage = function() {
       testUtil.sendKeys(fileInput, testFile);
       // Re-hide the file input element
       browser.executeScript("$('#fileAnchor')[0].className = 'hide'");
+    },
+
+
+    /**
+     *  Returns the QuestionnaireResponse (as an object) for the form on the page.
+     * @param options options for the getFormFHIRData call (the fourth
+     * parameter).
+     */
+    getQuestionnaireResponse: function(options) {
+      return browser.executeScript("return LForms.Util.getFormFHIRData("+
+       "'QuestionnaireResponse', getFHIRVersion(), null, arguments[0])", options);
     }
+
+
   });
   return rtnObj;
 };